Cleaned up and reorganized components
