+++ components/explore-the-data-page/CheckboxGroup.js
